Time: 12 ms (74.59%) | Memory: 20.4 MB (18.88%) - LeetSync

+++ 268-missing-number/missing-number.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int missingNumber(vector<int>& nums) {
-        long long c = nums[0] ^ nums.size();
+        int c = nums[0] ^ nums.size();
         for(int i=1;i<nums.size();i++) c  = c ^ nums[i] ^ i;
         return c;
     }
